# docs/openapi-spec.yaml
openapi: 3.0.3
info:
  title: NEXUS Platform API
  description: |
    # NEXUS Platform - Unified Backend API

    A comprehensive, AI-powered platform providing unified services for modern applications.

    ## Features

    * **SSOT (Single Source of Truth)**: Centralized configuration and alias management
    * **API Registry**: Service discovery and integration
    * **Rate Limiting**: Advanced rate limiting with multiple strategies
    * **Authentication**: JWT-based authentication with MFA support
    * **Financial Services**: Account management and transaction processing
    * **AI Services**: Chat, insights, and optimization features
    * **Monitoring**: Comprehensive system and performance monitoring
    * **Deployment**: Blue-green deployment and rollback capabilities

    ## Authentication

    Most endpoints require authentication. Include the JWT token in the Authorization header:

    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Rate Limiting

    API requests are rate-limited. Check the response headers for limit information:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests
    - `X-RateLimit-Reset`: Time when limit resets

    ## Versioning

    API endpoints support versioning. Use the version prefix in the URL:
    - `v1`: Legacy endpoints
    - `v3`: Current stable version

    ## Error Responses

    All error responses follow this format:

    ```json
    {
...
      "details": {}
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /api/v3/ssot/aliases:
    get:
      summary: List all aliases
      description: Retrieve all SSOT aliases
      tags:
        - SSOT
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of aliases
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    alias:
                      type: string
                    canonical_name:
                      type: string
                    created_at:
                      type: string
                      format: date-time
    post:
      summary: Create new alias
      description: Create a new SSOT alias
      tags:
        - SSOT
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - alias
                - canonical_name
              properties:
                alias:
                  type: string
                canonical_name:
                  type: string
                metadata:
                  type: object
      responses:
        '201':
          description: Alias created successfully
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v3/auth/login
...